tasks/theprojecto/app/web/requirements.txt
aiohttp
aiohttp_jinja2
pyjwt
